// gen/ts/src/nama/api/v1/setup_pb.d.ts
// @generated by protoc-gen-es v2.13.0 with parameter "target=js+dts,import_extension=js"
// @generated from file nama/api/v1/setup.proto (package nama.api.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file nama/api/v1/setup.proto.
 */
export declare const file_nama_api_v1_setup: GenFile;

/**
 * @generated from message nama.api.v1.Administrator
 */
export declare type Administrator = Message<"nama.api.v1.Administrator"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string display_name = 2;
   */
  displayName: string;

  /**
   * @generated from field: string email = 3;
   */
  email: string;
};

/**
 * Describes the message nama.api.v1.Administrator.
 * Use `create(AdministratorSchema)` to create a new message.
 */
export declare const AdministratorSchema: GenMessage<Administrator>;

/**
 * @generated from message nama.api.v1.GetStatusRequest
 */
export declare type GetStatusRequest = Message<"nama.api.v1.GetStatusRequest"> & {
};

/**
 * Describes the message nama.api.v1.GetStatusRequest.
 * Use `create(GetStatusRequestSchema)` to create a new message.
 */
export declare const GetStatusRequestSchema: GenMessage<GetStatusRequest>;

/**
 * @generated from message nama.api.v1.GetStatusResponse
 */
export declare type GetStatusResponse = Message<"nama.api.v1.GetStatusResponse"> & {
  /**
   * @generated from field: bool initialized = 1;
   */
  initialized: boolean;
};

/**
 * Describes the message nama.api.v1.GetStatusResponse.
 * Use `create(GetStatusResponseSchema)` to create a new message.
 */
export declare const GetStatusResponseSchema: GenMessage<GetStatusResponse>;

/**
 * @generated from message nama.api.v1.CreateAdministratorRequest
 */
export declare type CreateAdministratorRequest = Message<"nama.api.v1.CreateAdministratorRequest"> & {
  /**
   * @generated from field: string bootstrap_token = 1;
   */
  bootstrapToken: string;

  /**
   * @generated from field: string display_name = 2;
   */
  displayName: string;

  /**
   * @generated from field: string email = 3;
   */
  email: string;

  /**
   * @generated from field: string password = 4;
   */
  password: string;
};

/**
 * Describes the message nama.api.v1.CreateAdministratorRequest.
 * Use `create(CreateAdministratorRequestSchema)` to create a new message.
 */
export declare const CreateAdministratorRequestSchema: GenMessage<CreateAdministratorRequest>;

/**
 * @generated from message nama.api.v1.CreateAdministratorResponse
 */
export declare type CreateAdministratorResponse = Message<"nama.api.v1.CreateAdministratorResponse"> & {
  /**
   * @generated from field: nama.api.v1.Administrator administrator = 1;
   */
  administrator?: Administrator | undefined;
};

/**
 * Describes the message nama.api.v1.CreateAdministratorResponse.
 * Use `create(CreateAdministratorResponseSchema)` to create a new message.
 */
export declare const CreateAdministratorResponseSchema: GenMessage<CreateAdministratorResponse>;

/**
 * @generated from service nama.api.v1.SetupService
 */
export declare const SetupService: GenService<{
  /**
   * @generated from rpc nama.api.v1.SetupService.GetStatus
   */
  getStatus: {
    methodKind: "unary";
    input: typeof GetStatusRequestSchema;
    output: typeof GetStatusResponseSchema;
  },
  /**
   * @generated from rpc nama.api.v1.SetupService.CreateAdministrator
   */
  createAdministrator: {
    methodKind: "unary";
    input: typeof CreateAdministratorRequestSchema;
    output: typeof CreateAdministratorResponseSchema;
  },
}>;